// needrelay-frontend/src/api/generated/models/organizationRole.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type OrganizationRole = typeof OrganizationRole[keyof typeof OrganizationRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationRole = {
  ADMIN: 'ADMIN',
  USER: 'USER',
} as const;